requirements.txt
.
pyserial-asyncio-fast>=0.11
